+++ src/sanitizer.js
@@ -2,7 +2,14 @@
     ('use strict');
 
     const LOG = false;
-    const DEFAULT_SETTINGS = { banned: [], isPopularIgnored: false };
+    const DEFAULT_SETTINGS = { banned: [], isQuickActionsOn: false };
+    const selectors = {
+        author: '.user-info__nickname',
+        company: '.post__title a',
+        hub: '.hub-link',
+        article: 'article',
+        notHiddenArticle: `article:not(.sanitizer-hidden-article)`,
+    };
 
     /**
      * Logs message to console if log flag is true
@@ -10,6 +17,14 @@
      */
     function log(...message) {
         LOG && console.log(...message);
+    }
+
+    /**
+     * Check if string is empty (null, undefined, sapces)
+     * @param {string} v string value
+     */
+    function isEmpty(v) {
+        return v == null || v.toString().trim() === '';
     }
 
     /**
@@ -22,6 +37,19 @@
     }
 
     /**
+     * Checks if current page is personal
+     */
+    function onPersonalPage() {
+        const location = window.location.href.toLowerCase();
+
+        const onUsersPageRegex = /\/users\/.*\/posts/;
+        const onBlogPageRegex = /\/company\/.*\/blog/;
+        const onArticlePageRegex = /\/post\/d*/;
+
+        return onUsersPageRegex.test(location) || onBlogPageRegex.test(location) || onArticlePageRegex.test(location);
+    }
+
+    /**
      * Checks if article belongs to particular user
      * @param {HTMLElement} article
      * @returns {boolean}
@@ -29,7 +57,7 @@
     function belongsToAuthor(article, authorName) {
         return article == null
             ? false
-            : equalsCaseInsensetive(article.querySelector(`.user-info__nickname`)?.textContent, authorName);
+            : equalsCaseInsensetive(article.querySelector(selectors.author)?.textContent, authorName);
     }
 
     /**
@@ -40,7 +68,7 @@
     function belongsToBlog(article, blogName) {
         return article == null
             ? false
-            : article.querySelector(`.post__title a`)?.href?.toLowerCase()?.includes(`/company/${blogName}`);
+            : article.querySelector(selectors.company)?.href?.toLowerCase()?.includes(`/company/${blogName}`);
     }
 
     /**
@@ -49,13 +77,14 @@
      * @returns {boolean}
      */
     function belongsToHab(article, searchTerm) {
-        return [...(article?.querySelectorAll('.hub-link') ?? [])].some(
+        return [...(article?.querySelectorAll(selectors.hub) ?? [])].some(
             (el) => el.innerText?.toLowerCase() === searchTerm || el.href?.toLowerCase()?.includes(`/hub/${searchTerm}`)
         );
     }
 
     /**
      * Gets banned authors from the store
+     * @returns {Promise<{banned:Array<{name:string, disabled: boolean}>,isQuickActionsOn: boolean}>}
      */
     async function getSettings() {
         return new Promise((resolve, _) => {
@@ -67,68 +96,121 @@
     }
 
     /**
-     * Removes links from reading now block
-     * @param {number[]} ids
+     * Update settings
+     * @param {{banned:Array<{name:string, disabled: boolean}>,isQuickActionsOn: boolean}} Updates settings
      */
-    function removeLinksFromReadingNow(ids) {
-        const [...list] = document.querySelectorAll('#neuro-habr .content-list__item');
-        const banned = list.filter((x) => {
-            const href = (x.querySelector('.post-info__title')?.href ?? '').toLowerCase();
-            const containsBannedId = ids.some((id) => href.includes(`post/${id}`));
-            return containsBannedId;
-        });
-
-        banned.forEach((x) => (x.innerHTML = `<!-- removed-->`));
+    async function updateSettings(settings) {
+        return new Promise((resolve) => chrome.storage.sync.set({ settings }, () => resolve()));
     }
 
     /**
-     * Extracts article id from href
-     * @param {string} href
-     * @returns number
+     * Gets all visible (not banned) articles
      */
-    function extractArticleId(href) {
-        if (href == null) {
-            return NaN;
+    function getVisibleArticles() {
+        return [...document.querySelectorAll(selectors.notHiddenArticle)];
+    }
+
+    /**
+     * Update settings with new search term and hide articles belongs to search term
+     * @param {Event} e
+     */
+    async function hideBtnHandler(e) {
+        if (e.cancelable) {
+            e.preventDefault();
+            e.stopPropagation();
         }
 
-        const trimmed = href.endsWith('/') ? href.slice(0, -1) : href;
-        return trimmed.split('/').pop();
-    }
+        const rawSearchTerm = e.target.dataset.searchTerm;
 
-    function onPersonalPage(name) {
-        const location = window.location.href.toLowerCase();
-        const isOnAuthorPage = location.includes(`/users/${name}/posts`);
-        const isOnBlogPage = location.includes(`/company/${name}/blog`);
-        return isOnAuthorPage || isOnBlogPage;
+        if (isEmpty(rawSearchTerm)) {
+            return;
+        }
+
+        const searchTerm = rawSearchTerm.toLowerCase();
+        const settings = await getSettings();
+
+        const alreadyBanned = settings.banned.find((x) => equalsCaseInsensetive(x.name, searchTerm)); // in case of different pages
+
+        if (!alreadyBanned) {
+            settings.banned.push({ name: searchTerm, disabled: false });
+            await updateSettings(settings);
+        }
+
+        tryRemoveArticles(searchTerm, getVisibleArticles());
     }
 
     /**
-     * Gets id of the banned articles
-     * @param {HTMLElement[]} articles
-     * @param {string[]} banned list of banned authors/blogs
+     * Creates hide button
+     * @param {string} title button title
+     * @param {string} searchTerm key to ignore article
      */
-    function getIdForBannedArticles(articles, banned) {
-        return articles
-            .filter((article) =>
-                banned.some((searchTerm) => belongsToAuthor(article, searchTerm) || belongsToHab(article, searchTerm))
-            )
-            .map((article) => extractArticleId(article.querySelector('.post__title a')?.href));
+    function createHideBtn(title, searchTerm) {
+        const hideHubBtn = document.createElement('button');
+        hideHubBtn.className = 'sanitizer-action-remove-hub';
+        hideHubBtn.title = title;
+        hideHubBtn.innerHTML = `<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 572.098 572.098">
+        <path d="M99.187 398.999l44.333-44.332c-24.89-15.037-47.503-33.984-66.763-56.379 29.187-33.941 66.053-60.018 106.947-76.426-6.279 14.002-9.853 29.486-9.853 45.827 0 16.597 3.696 32.3 10.165 46.476l35.802-35.797c-5.698-5.594-9.248-13.36-9.248-21.977 0-17.02 13.801-30.82 30.82-30.82 8.611 0 16.383 3.55 21.971 9.248l32.534-32.534 36.635-36.628 18.366-18.373c-21.206-4.186-42.896-6.469-64.848-6.469-107.663 0-209.732 52.155-273.038 139.518L0 298.288l13.011 17.957c23.819 32.871 53.14 60.754 86.176 82.754zM459.208 188.998l-44.854 44.854c30.539 16.071 58.115 37.846 80.986 64.437-52.167 60.662-128.826 96.273-209.292 96.273-10.3 0-20.533-.6-30.661-1.744l-52.375 52.375c26.903 6.887 54.762 10.57 83.036 10.57 107.663 0 209.738-52.154 273.038-139.523l13.011-17.957-13.011-17.956c-27.063-37.332-61.242-68.177-99.878-91.329z"/>
+        <path d="M286.049 379.888c61.965 0 112.198-50.234 112.198-112.199 0-5.588-.545-11.035-1.335-16.402L269.647 378.56c5.368.789 10.814 1.328 16.402 1.328zM248.815 373.431L391.79 230.455l4.994-4.994 45.796-45.796 86.764-86.77c13.543-13.543 13.543-35.502 0-49.046-6.77-6.769-15.649-10.159-24.523-10.159s-17.754 3.384-24.522 10.159l-108.33 108.336-22.772 22.772-29.248 29.248-48.14 48.14-34.456 34.456-44.027 44.027-33.115 33.115-45.056 45.055-70.208 70.203c-13.543 13.543-13.543 35.502 0 49.045 6.769 6.77 15.649 10.16 24.523 10.16s17.754-3.385 24.523-10.16l88.899-88.898 50.086-50.086 15.837-15.831z"/></svg>`;
+        hideHubBtn.type = 'button';
+        hideHubBtn.dataset.searchTerm = searchTerm;
+        hideHubBtn.onclick = hideBtnHandler;
+
+        return hideHubBtn;
+    }
+
+    /**
+     * Add quick action to hide hub
+     * @param {HTMLElement} article
+     */
+    function addHideHubQuickAction(article) {
+        article.querySelectorAll('a.hub-link').forEach((a) => {
+            const searchTerm = a.innerText;
+            if (isEmpty(searchTerm)) {
+                return;
+            }
+
+            const hideHubBtn = createHideBtn('Add this hub to HabroSanitizer the banlist', searchTerm, hideBtnHandler);
+            a.insertAdjacentElement('afterend', hideHubBtn);
+        });
+    }
+
+    /**
+     *
+     * @param {HTMLElement} article
+     */
+    function addHideAuthorQuickAction(article) {
+        article.querySelectorAll('.user-info__nickname').forEach((t) => {
+            const searchTerm = t.innerText;
+            if (isEmpty(searchTerm)) {
+                return;
+            }
+
+            const hideAuthBtn = createHideBtn(`Add this author/blog the banlist`, searchTerm, hideBtnHandler);
+            t.insertAdjacentElement('afterend', hideAuthBtn);
+        });
+    }
+
+    /**
+     * Add remove action button (link) to each hub link of article
+     * @param {HTMLElement} article
+     */
+    function addQuickActionButtons(article) {
+        addHideHubQuickAction(article);
+        addHideAuthorQuickAction(article);
     }
 
     /**
      * Removes article from regular list by author name
-     * @param {string} author
+     * @param {string} keyword
      * @param {HTMLElement[]} articles
      */
-    function removeArticle(author, articles) {
-        const searchTerm = (author ?? '').toString().toLowerCase();
-
-        const isOnPersonalPage = onPersonalPage(author);
-
-        if (isOnPersonalPage) {
-            log(`We are on the ${searchTerm} personal page, skipping`);
+    function tryRemoveArticles(keyword, articles) {
+        if (isEmpty(keyword)) {
+            log(`Search term is null, ignoring`);
             return;
         }
+
+        const searchTerm = keyword.toString().toLowerCase();
 
         const articlesToBeDeleted = articles.filter(
             (article) =>
@@ -142,55 +224,20 @@
         articlesToBeDeleted.forEach((article) => article.classList.add('sanitizer-hidden-article'));
     }
 
-    /**
-     * Gets all visible (not banned) articles
-     */
-    function get_visible_articles() {
-        return [...document.querySelectorAll('article:not(.post_full):not(.sanitizer-hidden-article)')];
-    }
+    const isOnPersonalPage = onPersonalPage();
 
-    /**
-     * Add remove action button (link) to each hub link of article
-     * @param {HTMLElement[]} article
-     */
-    function addContentActions(article) {
-        article.querySelectorAll('a.hub-link').forEach(a => {
-            const remove_link = document.createElement('button');
-            remove_link.className = 'sanitizer-action-remove-hub';
-            remove_link.title = 'Add this hub to HabroSanitizer banned list';
-            remove_link.innerHTML = '&times;';
-            remove_link.onclick = addHubToBanned;
-            a.insertAdjacentElement('afterend', remove_link);
-        });
-    }
-
-    /**
-     * Event handler for append hub name to ban list and hide all related visible articles
-     * @param {Event[]} click event
-     */
-    function addHubToBanned(event) {
-        const a = event.currentTarget.previousElementSibling, hubname = a.textContent.toLocaleLowerCase();
-        settings.banned.push({ name: hubname, disabled: false });
-        chrome.storage.sync.set({ settings }, () => removeArticle(hubname, get_visible_articles()));
+    if (isOnPersonalPage) {
+        log('We are on the perosnal page, HabroSanitizer actions disabled');
+        return;
     }
 
     const settings = await getSettings();
     const banned = settings.banned.map(({ name }) => name);
-    const isReadingNowBlockOff = settings.isPopularIgnored && false; // see #6
-
     log(`Found list of banned users: ${banned.join(',')} `);
+    const [...allArticles] = document.querySelectorAll(selectors.article);
 
-    const [...allArticles] = document.querySelectorAll('article:not(.post_full)');
-
-    if (isReadingNowBlockOff) {
-        const listOfBannedId = getIdForBannedArticles(allArticles, banned);
-        console.log(`Found ids for ban: "${listOfBannedId.join(',')}"`);
-        removeLinksFromReadingNow(listOfBannedId);
-    }
-
-    banned.forEach((name) => removeArticle(name, allArticles));
-    if ( settings.show_hub_actions )
-        get_visible_articles().forEach(addContentActions); // add hub action buttons for all visible articles
+    banned.forEach((name) => tryRemoveArticles(name, allArticles));
+    settings.isQuickActionsOn && getVisibleArticles().forEach(addQuickActionButtons);
 
     log(`Sanitization done`);
 })();